requirements.txt
Flask==2.0.1
flask-cors==3.0.10
gunicorn==20.1.0
python-dotenv==0.19.0
requests==2.27.0
Werkzeug==2.0.1